LayoutTests/fast/dom/window-xy-properties-expected.txt

EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document toDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document toDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 50
PASS event.pageY is 50
PASS event.layerX is 50
PASS event.layerY is 50
PASS event.offsetX is 42
PASS event.offsetY is 42
PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 50
PASS event.pageY is 550
PASS event.layerX is 50
PASS event.layerY is 550
PASS event.offsetX is 42
PASS event.offsetY is 542
PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 550
PASS event.pageY is 50
PASS event.layerX is 550
PASS event.layerY is 50
PASS event.offsetX is 542
PASS event.offsetY is 42
PASS ev.clientX is 50
PASS ev.clientY is 50
PASS ev.pageX is 550
PASS ev.pageY is 550
PASS ev.layerX is 550
PASS ev.layerY is 550
PASS ev.offsetX is 542
PASS ev.offsetY is 542
